Initialize the event manager when creating the activity.
This is maybe only temporary; the event manager hits the disk (i.e. creates/gets the database) in init, so it could be good to do in a background thread. On the other hand, a synchronous API would be way better, and this only happens in launch. All other event database handling should probably happen in the background, though.