Methods for creating and getting budgets, with corresponding test.
For now let’s get all budgets at the same time. They’re not big, and should be deleted anyway after a while. Speaking of deletion: it’s pretty hard to do now, since there’s no connection with the budgets or events to the ones in the database. Probably gotta put the row IDs in the models.