Bumped the minimum SDK version to 21.
This isn’t related to this branch, but I did remember something important: this app is for me, not all Android users. My current phone is on Marshmallow, so no reason for me to support older versions of Android. I don’t have any older versions available to me either.